# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DepartmentProfile {
  name: string;
  institution: string;
  shortName: string;
  location: string;
  established: string;
  description: string;
}

export type DashboardCounts = {
  students: number;
  faculty: number;
  upcomingEvents: number;
  unreadNotifications: number;
};

export interface Highlight {
  label: string;
  value: string;
  detail: string;
  tone: string;
}

export interface Event {
  id: number;
  title: string;
  date: string;
  time: string;
  venue: string;
  type: string;
  description: string;
}

export interface NewsItem {
  id: number;
  title: string;
  excerpt: string;
  publishedAt: string;
  category: string;
  featured: boolean;
}

export interface Dashboard {
  profile: DepartmentProfile;
  counts: DashboardCounts;
  highlights: Highlight[];
  upcomingEvents: Event[];
  latestNews: NewsItem[];
}

export interface Student {
  id: number;
  name: string;
  rollNumber: string;
  year: string;
  semester: string;
  email: string;
  focus: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface Faculty {
  id: number;
  name: string;
  designation: string;
  specialization: string;
  email: string;
  office: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface Notification {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  category: string;
  read: boolean;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface SessionUser {
  name: string;
  email: string;
  role: string;
  initials: string;
}

export interface Session {
  authenticated: boolean;
  user: SessionUser | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetStudentsParams = {
year?: string;
search?: string;
};

export type GetEventsParams = {
upcoming?: boolean;
};