# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardCounts } from './dashboardCounts';
import type { DepartmentProfile } from './departmentProfile';
import type { Event } from './event';
import type { Highlight } from './highlight';
import type { NewsItem } from './newsItem';

export interface Dashboard {
  profile: DepartmentProfile;
  counts: DashboardCounts;
  highlights: Highlight[];
  upcomingEvents: Event[];
  latestNews: NewsItem[];
}