# lib/api-zod/src/generated/types/departmentProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentProfile {
  name: string;
  institution: string;
  shortName: string;
  location: string;
  established: string;
  description: string;
}