# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  date: string;
  time: string;
  venue: string;
  type: string;
  description: string;
}