# lib/api-zod/src/generated/types/faculty.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Faculty {
  id: number;
  name: string;
  designation: string;
  specialization: string;
  email: string;
  office: string;
  /** @nullable */
  avatarUrl: string | null;
}