# lib/api-zod/src/generated/types/highlight.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Highlight {
  label: string;
  value: string;
  detail: string;
  tone: string;
}