# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  title: string;
  excerpt: string;
  publishedAt: string;
  category: string;
  featured: boolean;
}