# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  category: string;
  read: boolean;
}