# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionUser } from './sessionUser';

export interface Session {
  authenticated: boolean;
  user: SessionUser | null;
}