# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  rollNumber: string;
  year: string;
  semester: string;
  email: string;
  focus: string;
  /** @nullable */
  avatarUrl: string | null;
}